Visual part of both registration and login
Various options have been added to the application's main page, allowing users to both register and log in. Additionally, there are validations for different fields (e.g., email, password, etc.) and a field where the user will be asked for their email verification code. Logging in redirects to one of the page's menus. However, there is no design; that remains to be seen.